CourtCaseManagementSystem/src/Controller/DatabaseCntl/CCMSDatabaseStubTest.java


package Controller.DatabaseCntl;

        import Controller.Exception.CourtCaseMgmtSystemException;
        import Controller.CivilianCntl.Civilian;
        import org.junit.jupiter.api.BeforeEach;
        import org.junit.jupiter.api.Test;

        import static org.junit.jupiter.api.Assertions.*;

class CCMSDatabaseStubTest {

    private Civilian civilian;
    private CCMSDatabaseStub databaseStub;

    @BeforeEach
    void setUp() throws CourtCaseMgmtSystemException {
        databaseStub = new CCMSDatabaseStub("testUser", "testPassword");
        civilian = new Civilian("testUser", "testPassword");
    }

    @Test
    void testUserAuthorization() {
        assertTrue(databaseStub.isUserAuthorized("testUser", "testPassword"));
    }

    @Test
    void testDatabaseConnection() {
        assertTrue(databaseStub.connectToCCMSDatabase());
    }

    @Test
    void testDatabaseManagementSchedule() {
        assertTrue(databaseStub.databaseManagementSchedule());
    }

    @Test
    void testManualTrigger() {
        assertTrue(databaseStub.manuallyTriggerByTheCourtAdministrator());
    }

    // ... continue with other methods ...

    // If you have methods that interact with the Civilian class, test those interactions too
    @Test
    void testCivilianDatabaseInteraction() throws CourtCaseMgmtSystemException {
        // Given some state or setup

        // When a method in Civilian is called that interacts with the database
        boolean result = civilian.searchSpecificCase();

        // Then the expected outcome should be true (or whatever you expect)
        assertTrue(result);
    }

    // ... continue with other interactions ...
}
